+++ EnsNetcode/Netcode/Common/ProtocolWrapper/Base/ListenerBase.cs
@@ -8,7 +8,7 @@ namespace ProtocolWrapper
     {
         protected IPAddress IP;
         protected int Port;
-        protected int connectionIndex=0;
+        protected short connectionIndex=0;
         public bool Listening=false;
         public bool Cancelled=false;
